Organize runtime value inspection
- Split compound value formatters into semantic owner modules.
- Keep value dispatch and primitive formatting in the inspection facade.
- Place exact output tests beside each compound formatter.
